Devolvemos las peliculas con su clasificacion "real"

+++ backend-django/api/models.py
@@ -10,7 +10,7 @@ class Pelicula(models.Model):
     director = models.CharField(max_length=100)
     poster_url = models.URLField(max_length=200)    # imagen
     total_votos = models.PositiveIntegerField(default=0)
-    promedio_votos = models.DecimalField(max_digits=3, decimal_places=2, default=0)
+    promedio_votos = models.FloatField(default=0)
     tmdb_id = models.PositiveIntegerField(unique=True)
 
     def __str__(self):
